fix: refactored types for general purposes
