+++ src/App/category.resolver.ts
@@ -12,7 +12,7 @@ import { MinLength } from 'class-validator'
 import Category from '../Database/Entities/category.entity'
 import PaginationArgs from './pagination.args'
 import Wiki from '../Database/Entities/wiki.entity'
-import { ICategory } from '../Database/Entities/types/ICategory'
+import { ICategory } from '../types/ICategory'
 
 @ArgsType()
 class TitleArgs {
