--- src/Database/Entities/types/IWiki.ts
@@ -1,7 +1,0 @@
-// TODO: add all types
-// right now only used ones are added
-
-export interface IWiki {
-  id: number
-  title: string
-}
