+++ src/Indexer/Validator/validator.service.ts
@@ -3,38 +3,38 @@ import { ConfigService } from '@nestjs/config'
 
 import * as linkify from 'linkifyjs'
 
-import { ValidWiki } from '../Store/store.service'
+import { IWiki } from '../../types/IWiki'
 
 @Injectable()
 class IPFSValidatorService {
   private configService: ConfigService = new ConfigService()
 
   async validate(
-    wiki: ValidWiki,
+    wiki: IWiki,
     validateJSON?: boolean,
     hashUserId?: string,
   ): Promise<boolean> {
     const languages = ['en', 'es', 'ko', 'zh']
 
-    const checkLanguage = (validatingWiki: ValidWiki) =>
+    const checkLanguage = (validatingWiki: IWiki) =>
       !!languages.includes(validatingWiki.language)
 
-    const checkWords = (validatingWiki: ValidWiki) =>
+    const checkWords = (validatingWiki: IWiki) =>
       validatingWiki.content.split(' ').length >= 150
 
-    const checkMetadataChars = (validatingWiki: ValidWiki) =>
+    const checkMetadataChars = (validatingWiki: IWiki) =>
       validatingWiki.metadata[1]?.value.length < 255
 
-    const checkCategories = (validatingWiki: ValidWiki) =>
+    const checkCategories = (validatingWiki: IWiki) =>
       validatingWiki.categories.length === 1
 
-    const checkUser = (validatingWiki: ValidWiki) => {
+    const checkUser = (validatingWiki: IWiki) => {
       if (validateJSON) return true
 
       return validatingWiki.user.id.toLowerCase() === hashUserId?.toLowerCase()
     }
 
-    const checkSummary = (validatingWiki: ValidWiki) => {
+    const checkSummary = (validatingWiki: IWiki) => {
       if (validatingWiki.summary) {
         return validatingWiki.summary.length <= 128
       }
@@ -42,7 +42,7 @@ class IPFSValidatorService {
       return true
     }
 
-    const checkImages = (validatingWiki: ValidWiki) => {
+    const checkImages = (validatingWiki: IWiki) => {
       if (
         validatingWiki.images.length >= 1 &&
         validatingWiki.images.length <= 5
@@ -53,7 +53,7 @@ class IPFSValidatorService {
       return false
     }
 
-    const checkExternalUrls = (validatingWiki: ValidWiki) => {
+    const checkExternalUrls = (validatingWiki: IWiki) => {
       const links = linkify.find(validatingWiki.content)
       const uiLink = this.configService.get('UI_URL')
 
@@ -66,7 +66,7 @@ class IPFSValidatorService {
       return false
     }
 
-    const checkTags = (validatingWiki: ValidWiki) =>
+    const checkTags = (validatingWiki: IWiki) =>
       validatingWiki.tags.length <= 5
 
     console.log('🕦 Validating Wiki content from IPFS 🕦')
